# backend/docs/swagger.yaml
basePath: /
definitions:
  models.BuyTokensResponse:
    properties:
      message:
        type: string
    type: object
...
  models.DistributeTokensResponse:
    properties:
      message:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.Kermesse:
    properties:
      created_at:
        type: string
      description:
        type: string
      end_date:
        type: string
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      organisateurs:
        description: Relation many-to-many avec les organisateurs
        items:
          $ref: '#/definitions/models.User'
        type: array
      participants:
        description: Relation many-to-many avec les participants
        items:
          $ref: '#/definitions/models.User'
        type: array
      stands:
        description: Association one-to-many avec les stands
        items:
          $ref: '#/definitions/models.Stand'
        type: array
      start_date:
        type: string
      updated_at:
        type: string
    type: object
  models.KermesseStats:
    properties:
      created_at:
        type: string
      id:
        type: integer
      kermesse_id:
        type: integer
      total_participants:
        type: integer
      total_revenue:
        type: number
      total_tokens_used:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.OrganisateurStats:
    properties:
      created_at:
        type: string
      id:
        type: integer
      organisateur_id:
        type: integer
      total_kermesses:
        type: integer
      total_revenue:
        type: number
      updated_at:
        type: string
    type: object
...
  models.Stand:
    properties:
      created_at:
        description: Date de création du stand
        type: string
      description:
        description: Augmenté la taille pour plus de flexibilité
        type: string
      end_date:
        description: Date de fin du stand
        type: string
      id:
        type: integer
      kermesse:
        allOf:
        - $ref: '#/definitions/models.Kermesse'
        description: Associations
      kermesse_id:
        description: Référence vers la Kermesse
        type: integer
      location:
        description: Lieu où se trouve le stand
        type: string
      name:
        type: string
      owner:
        $ref: '#/definitions/models.User'
      owner_id:
        description: Référence vers le User qui tient le stand
        type: integer
      start_date:
        description: Date de début du stand
        type: string
      stock:
        type: integer
      type:
        description: e.g., "food", "game"
        type: string
      updated_at:
        description: Date de la dernière mise à jour du stand
        type: string
    type: object
  models.StandStats:
    properties:
      created_at:
        type: string
      id:
        type: integer
      items_sold:
        type: integer
      revenue:
        type: number
      stand_id:
        type: integer
      total_tokens_used:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      passwordHash:
        type: string
      resetToken:
        description: Token de réinitialisation
        type: string
      resetTokenExpiry:
        description: Expiration du token de réinitialisation
        type: string
      role:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: contact@dratsisama.com
    name: Dr_AtsiSama
    url: http://dratsisama.com
  description: Documentation de l'API pour la gestion des kermesses scolaires avec
    fonctionnalités de stand, tombola, utilisateurs, etc.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  title: Kermisys API Documentation
  version: "1.0"
paths:
...
  /stands:
    get:
      description: Récupère la liste de tous les stands avec pagination
      parameters:
      - default: 1
        description: Numéro de la page
        in: query
        name: page
        type: integer
      - default: 10
        description: Nombre d'éléments par page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Stand'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Récupérer tous les stands avec pagination
      tags:
      - Stands
    post:
      consumes:
      - application/json
      description: Ajoute un nouveau stand
      parameters:
      - description: Détails du stand (Nom, Description, etc.)
        in: body
        name: stand
        required: true
        schema:
          $ref: '#/definitions/models.Stand'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Stand'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Créer un stand
      tags:
      - Stands
  /stands/{id}:
    delete:
      description: Supprime un stand par son ID
      parameters:
      - description: ID du stand
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Stand deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Supprimer un stand
      tags:
      - Stands
    get:
      description: Récupère les détails d'un stand par son ID
      parameters:
      - description: ID du stand
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Stand'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Récupérer un stand par ID
      tags:
      - Stands
    put:
      consumes:
      - application/json
      description: Met à jour les détails d'un stand existant
      parameters:
      - description: ID du stand
        in: path
        name: id
        required: true
        type: integer
      - description: Détails du stand
        in: body
        name: stand
        required: true
        schema:
          $ref: '#/definitions/models.Stand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Stand'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Mettre à jour un stand
      tags:
      - Stands
  /stands/{id}/interact:
    post:
      description: Permet à un utilisateur d'interagir avec un stand spécifique
      parameters:
      - description: ID du stand
        in: path
        name: id
        required: true
        type: integer
      - description: 'Action effectuée (ex: ''buy_item'', ''play_game'')'
        in: query
        name: action
        required: true
        type: string
      - description: Quantité à affecter (par défaut 1)
        in: query
        name: quantity
        type: integer
      responses:
        "200":
          description: Interaction réussie
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Interagir avec un stand
      tags:
      - Stands
  /stats/kermesse/{id}:
    get:
      description: Récupère les statistiques de la kermesse par ID
      parameters:
      - description: ID de la kermesse
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.KermesseStats'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Obtenir les statistiques de la kermesse
      tags:
      - Statistiques
  /stats/organisateur/{id}:
    get:
      description: Récupère les statistiques de l'organisateur par ID
      parameters:
      - description: ID de l'organisateur
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OrganisateurStats'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Obtenir les statistiques de l'organisateur
      tags:
      - Statistiques
  /stats/stand/{id}:
    get:
      description: Récupère les statistiques d'un stand par ID
      parameters:
      - description: ID du stand
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StandStats'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Obtenir les statistiques d'un stand
      tags:
      - Statistiques
  /tokens/buy:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Permet à un utilisateur d'acheter des jetons
      parameters:
      - description: Nom d'utilisateur
        in: header
        name: username
        required: true
        type: string
      - description: Nombre de jetons à acheter
        in: formData
        name: amount
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BuyTokensResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Achat de jetons
      tags:
      - Jetons
  /tokens/distribute:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Permet à un parent de distribuer des jetons à son enfant
      parameters:
      - description: Nom d'utilisateur du parent
        in: header
        name: username
        required: true
        type: string
      - description: Nom d'utilisateur de l'enfant
        in: formData
        name: child
        required: true
        type: string
      - description: Nombre de jetons à distribuer
        in: formData
        name: amount
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DistributeTokensResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Distribution de jetons
      tags:
      - Jetons
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"